+++ tests/pds/tf_sheriff/__init__.py
@@ -1,2 +1,0 @@
-# encoding: utf-8
-"""Tests for tf_sheriff."""